+++ Examples/StronglyTypedData.ps1
@@ -1,28 +1,26 @@
 
 <#
 .Synopsis
-	Exports and imports instances of PowerShell classes.
+	Exports and imports using PowerShell classes and psd1 files.
 
 .Description
-	This scripts demonstrates strongly typed data logging using psd1 files.
-	Schema:
+	This scripts demonstrates strongly typed data logging using psd1 files:
 	- Define required PowerShell classes, e.g. [Message] and related types.
-	- Writing: use ConvertTo-Psd with Depth + Set-Content or Add-Content.
-	- Reading: use Import-Psd and convert raw data to [Message[]].
+	- To export, use ConvertTo-Psd with Depth + Set-Content or Add-Content.
+	- To import, use Import-Psd + convert raw data to [Message[]].
 
-	We create two messages and write them one by one to the file z.psd1.
-	The first one creates the file, the second is appended to it. Then
-	we read the messages, convert to the original type, and output:
+	We create two messages and write them one by one to the file "z.psd1".
+	The first creates the file, the second is appended to it. Then we read
+	data, convert to the original type, and output [Message] objects:
 
-		Message                Severity Location
-		-------                -------- --------
-		Obsolete feature.       Warning {C:\Scripts\Script1.ps1:11}
-		Not supported feature.    Error {C:\Scripts\Script2.ps1:99}
+	Message                Severity Location
+	-------                -------- --------
+	Obsolete feature.       Warning {C:\Scripts\Script1.ps1:11}
+	Not supported feature.    Error {C:\Scripts\Script2.ps1:99}
 #>
-
 #requires -Version 5
 
-# Define the message data types.
+# Define data types.
 enum Severity {
 	Information
 	Warning
@@ -39,7 +37,7 @@ class Message {
 	[Location] $Location
 }
 
-# Create a message and save it to the new file.
+# Create message 1 and save it to the new file.
 [Message] @{
 	Message = 'Obsolete feature.'
 	Severity = 'Warning'
@@ -47,7 +45,7 @@ class Message {
 } |
 ConvertTo-Psd -Depth 2 | Set-Content z.psd1
 
-# Create another message and append it to the file.
+# Create message 2 and append it to the file.
 [Message] @{
 	Message = 'Not supported feature.'
 	Severity = 'Error'