+++ Tests/Convert-XmlToPsd.test.ps1
@@ -4,5 +4,5 @@
 task BadXml {
 	$xml = [xml]'<bar><bad/></bar>'
 	($r = try {Convert-XmlToPsd $xml} catch {$_})
-	equals "$r" "Unexpected XML element 'bad'."
+	equals "$r" "Unexpected node 'bad'."
 }