+++ Tests/Export-PsdXml.test.ps1
@@ -6,6 +6,6 @@ task BadXml {
 	Remove-Item [z].psd1
 	$xml = [xml]'<bar><bad/></bar>'
 	($r = try {Export-PsdXml z.psd1 $xml} catch {$_})
-	equals "$r" "Unexpected XML element 'bad'."
+	equals "$r" "Unexpected node 'bad'."
 	assert (!(Test-Path z.psd1))
 }